feat(cats): dbc wears the host's colors
A dbc pane painted its own muted green next to a cats session in
whatever theme the user actually chose, which read as a foreign
application pasted into the window.

The mapping is small because the two palettes are related — dbc's
surfaces were ported from cdx and cats' theme system descends from the
same place, so nine of the ten are a rename. Only the selection is
computed, and it is computed the way cats computes its own (accent over
background at 0.30), because the host's own selection key arrives as
rgba() and a terminal cell is either one color or another.

HEX ONLY, ALL OR NOTHING: a missing or non-hex value in any of the seven
core keys abandons the whole synthesis rather than shipping a palette
that is half the host's and half dbc's. A half-translated theme looks
like a bug, and the built-in palette is a perfectly good answer.

The startup fetch is the one synchronous control call dbc makes, bounded
by the probe timeout and made before build(), because tview primitives
copy the theme when they are CONSTRUCTED — fetching on a goroutine
instead would paint one frame in the wrong palette and then repaint, a
visible flash on every launch to save milliseconds nobody can perceive.
Live changes arrive over the event stream, whose filter names events and
never a pane: theme_changed is session-scoped against pane 0, so
narrowing to our own pane would silently lose it.

Making the palette a runtime value turned the color tags from constants
into vars, which had two consequences worth naming. vet now reads a
concatenated tag as a non-constant format string, so the seven calls
that passed a finished message to logf go to a new log() instead. And
two places kept colors that no longer followed: the layout containers
between the panes, and — the interesting one — a TextView's SECOND
style for its text area, captured from tview.Styles at construction,
which pane() never reaches because it takes a *Box. The log pane's
interior had therefore always been the deep surface rather than the
panel its border wears; it is set explicitly now to exactly that color,
which changes nothing about how it looks and everything about whether it
follows a theme.

Found while writing the test: QueueUpdateDraw runs the draw AFTER the
closure returns, so a test that reads cells when its closure finishes is
looking at the frame before the change.

+++ ui/modals.go
@@ -25,7 +25,7 @@ func center(p tview.Primitive, w, h int) tview.Primitive {
 
 func (a *App) showExportModal() {
 	if a.lastRes == nil {
-		a.logf(tagWarn + "no result to export — run a query first")
+		a.log(tagWarn + "no result to export — run a query first")
 		return
 	}
 	var form *tview.Form
@@ -72,7 +72,7 @@ func (a *App) showExportModal() {
 func (a *App) showHistoryModal() {
 	entries := a.hist.recent()
 	if len(entries) == 0 {
-		a.logf(tagWarn + "no query history yet — it fills up as you run queries")
+		a.log(tagWarn + "no query history yet — it fills up as you run queries")
 		return
 	}
